start pulling images endpoints into a feature dir
