+++ src/features/images/images.controllers.ts
@@ -1,0 +1,11 @@
+import { getImageList } from './images.service';
+import type { RequestHandler } from 'express';
+
+export const listImages: RequestHandler = async (req, res, next) => {
+  try {
+    const result = await getImageList();
+    res.status(201).json({ images: result });
+  } catch (error) {
+    next(error);
+  }
+};
