+++ src/features/images/images.routes.ts
@@ -1,0 +1,6 @@
+import { Router } from 'express';
+import { listImages } from './images.controllers';
+
+export const imagesRoutes = Router();
+
+imagesRoutes.get('list', listImages);
