+++ src/features/images/images.types.ts
@@ -1,0 +1,4 @@
+export interface ImageData {
+  fileName: string;
+  url: string;
+}
