+++ src/shared/aws.ts
@@ -1,0 +1,10 @@
+import { S3Client } from '@aws-sdk/client-s3';
+import { config } from './config';
+
+export const s3 = new S3Client({
+  region: 'us-east-1',
+  credentials: {
+    accessKeyId: config.awsAccessKeyId,
+    secretAccessKey: config.awsSecretAccessKey,
+  },
+});
